Rename folder to panels.
